requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.0.2
psycopg2-binary==2.8.6
pycparser==2.20
python-lichess==0.9
requests==2.25.0
six==1.15.0
slackclient==2.9.3
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
yarl==1.6.3
